requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
dataclasses==0.7
Django==2.2.13
djangorestframework==3.11.1
flake8==3.8.3
importlib-metadata==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.9.27
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.2.0
